requirements.txt
Django==5.2.6
djangorestframework==3.16.1
psycopg2-binary==2.9.10
python-decouple==3.8
# django-cors-headers==4.8.0
# djangorestframework-simplejwt==5.3.1
